Attempt to fix DockerHub Autobuild
The test after the build failed since the sdd source was not mounted
into the container, and hence the test script not available.

The extension of the test/run.sh script can be locally tested from the
source code root:

    docker run --rm -t -v $(pwd):/opt/sdd/ pylipp/sdd:latest /opt/sdd/test/run.sh

+++ test/run.sh
@@ -9,6 +9,13 @@
 # Environment variables:
 #   NO_APP_TESTS    If non-empty, run only framework tests
 
+if ! command -v docker &>/dev/null; then
+    # Assume container environment; e.g. in the context of DockerHub Autobuild/test
+    cd /opt/sdd/test  # workaround for relative framework/ paths in tests
+    bats -r framework
+    exit $?
+fi
+
 container_id=$(docker run -d -it --rm --volume "$(pwd)":/opt/sdd --workdir /opt/sdd/test pylipp/sdd:latest)
 
 if [ "$1" = "--open" ]; then
